Step 4: Adding Visibility Filter
This step added a feature to filter todo items by one of the following filter types:
 - SHOW_ALL
 - SHOW_ACTIVE
 - SHOW_COMPLETED

The original state was only an array containing todo items but now we also need to store
the selected filter. To do this, I created a new root reducer that returns the state
as an object. The original array state, as well as the new state for visibility filter,
are now properties of the root state.

This pattern is also known as "Reducer Composition with Object".

In the TodosApp component, I added a footer of links so filters can be selected. Also, the
displayed todo items are filtered by the selected filter type.

A corresponding action creator for setting visibility filter was also added.