Step 6: Containers with connect() from React Redux
One issue in the previous implementation is that some state properties need
to be passed from TodosApp down to child components. This can be a problem
in large applications. Also, we need to explicitly pass Redux store
from index.js to TodosApp.

In this step, we solve this problem by creating intermediate containers
with connect() from React Redux.