Step 9: Inject params using withRouter()
In this step I moved the logic to filter visible todo items from the TodoApp component
to the TodoList component.

Also I made further refactoring to inject visibility filter to TodoApp so that it does
not have to be passed from TodoApp component (as the TodoApp does not really use it).