tep 13 - Fake backend and fetching data with action
There are two major changes in this step. First, local storage is replaced with a fake backend
to simulate asynchronous requests to get todos from API.

The second major change is to create a new middleware to support Promise as an action. This allows
me to explore how we can wrap multiple middlewares around the raw dispatch function in Redux store,
which will eventually become applyMiddleware().